Make no_std; tighten bounds on ChdHasher::Hash
Dependency on `std` has been removed, since it is not needed. Accordingly, the `rand` crate's `std` feature has been removed as well.

The bounds on the `ChdHasher` trait's associated type `Hash` have become stricter. This is for correctness reasons: while generating a PHF, `Hash` needs to be able to losslessly convert to `usize` at some places. The `usize_cast::IntoUsize` trait enforces this.